perf: Refatorando o código para melhor organização

+++ controllers/userController.js
@@ -1,19 +1,18 @@
 import * as userModel from "../models/User.js";
 
-
 export const getById = async (req, res) => {
   try {
     const { id } = req.params;
     const user = await userModel.getUserById(id);
 
     if (!user) {
-      res.status(404).json({ msg: "Usuário não encontrado!" });
+      res.status(404).json({ error: "Nenhum usuário encontrado com esse ID!" });
     }
 
     res.status(200).json(user);
   } catch (error) {
     console.error(error);
-    res.status(500).json({ error: "Erro interno do servidor" });
+    res.status(500).json({ error: "Erro interno do servidor." });
   }
 };
 
@@ -21,14 +20,14 @@ export const getAllUsers = async (req, res) => {
   try {
     const users = await userModel.getAllUsers();
 
-    if (!users) {
-      res.status(404).json({ msg: "Nenhum usuário encontrado" });
+    if (users.length === 0) {
+      return res.status(204).json({ msg: "Nenhum usuário encontrado!" });
     }
 
     res.status(200).json(users);
   } catch (error) {
-    console.error("Erro na requisição do usuário", error);
-    res.status(500).json({ error: "Erro interno do servidor" });
+    console.error("Erro na requisição do usuário: ", error);
+    res.status(500).json({ error: "Erro interno do servidor." });
   }
 };
 
@@ -36,12 +35,25 @@ export const createUser = async (req, res) => {
   try {
     const { username, password, name, email, department, roles } = req.body;
 
-    const existingUser = await userModel.getUserByUsername(username);
-    if (existingUser) {
-      return res.status(400).json({ error: "Usuário já existente." });
+    const userExists = await userModel.getUserByUsername(username);
+    if (userExists) {
+      return res.status(400).json({
+        error:
+          "Usuário já cadastrado. Por favor, escolha outro nome de usuário.",
+      });
     }
 
-    const response = {
+    if (!username || !password || !name || !email || !department || !roles) {
+      return res.status(400).json({ error: "Campos obrigatórios ausentes." });
+    }
+
+    if (password.length < 8) {
+      res
+        .status(400)
+        .json({ error: "A senha precisa ter no mínimo 8 caracteres." });
+    }
+
+    const userData = {
       username,
       password,
       name,
@@ -50,13 +62,7 @@ export const createUser = async (req, res) => {
       roles,
     };
 
-    if (password.length < 8) {
-      res
-        .status(400)
-        .json({ error: "A senha precisa ter no mínimo 8 caracteres." });
-    }
-
-    await userModel.createUser(response);
+    await userModel.createUser(userData);
 
     res.status(201).json({ msg: "Usuário criado com sucesso!" });
   } catch (error) {
@@ -70,6 +76,19 @@ export const updateUser = async (req, res) => {
     const { username, name, email, password, department, roles, active } =
       req.body;
     const { id } = req.params;
+
+    const userExists = await userModel.getUserById(id);
+    if (!userExists) {
+      return res
+        .status(404)
+        .json({ error: "Nenhum usuário encontrado com esse ID!" });
+    }
+
+    if (password < 8) {
+      return res
+        .status(400)
+        .json({ error: "A senha precisa ter no mínimo 8 caracteres." });
+    }
 
     const response = {
       username,
@@ -86,7 +105,7 @@ export const updateUser = async (req, res) => {
     res.status(200).json({ msg: "Usuário atualizado com sucesso!" });
   } catch (error) {
     console.error(error);
-    res.status(500).json({ error: "Erro interno do servidor" });
+    res.status(500).json({ error: "Erro interno do servidor." });
   }
 };
 
@@ -94,11 +113,18 @@ export const deleteUser = async (req, res) => {
   try {
     const { id } = req.params;
 
+    const userExists = await userModel.getUserById(id);
+    if (!userExists) {
+      return res
+        .status(404)
+        .json({ error: "Nenhum usuário encontrado com esse ID!" });
+    }
+
     await userModel.deleteUser(id);
 
     res.status(200).json({ msg: "Usuário excluído com sucesso!" });
   } catch (error) {
     console.error(error);
-    res.status(500).json({ error: "Erro interno do servidor" });
+    res.status(500).json({ error: "Erro interno do servidor." });
   }
 };
